+++ packages/renderer/src/bootstrap/index.ts
@@ -6,3 +6,6 @@
 
 export { DefaultGameBootstrap } from './DefaultGameBootstrap'
 export type { GameBootstrapOptions } from './DefaultGameBootstrap'
+export type { MarioGameBootstrap } from './MarioGameBootstrap'
+export type { MarioGameBootstrapOptions } from './MarioGameBootstrap'
+export { DefaultMarioGameBootstrap } from './MarioGameBootstrap'